remove builder pattern
