+++ src/lib.rs
@@ -37,7 +37,6 @@ mod audit;
 mod crypto;
 mod event;
 mod feed;
-mod feed_builder;
 mod proof;
 mod replicate;
 mod storage;
@@ -45,7 +44,6 @@ mod storage;
 pub use crate::crypto::{generate_keypair, sign, verify, Signature};
 pub use crate::event::Event;
 pub use crate::feed::Feed;
-pub use crate::feed_builder::FeedBuilder;
 pub use crate::proof::Proof;
 pub use crate::replicate::Peer;
 pub use crate::storage::{Node, NodeTrait, Storage, Store};
